backend/requirements.txt
# App
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Agno
agno>=2.0.0
openai>=1.0.0

# QuickBooks
intuit-oauth==1.2.6
requests==2.32.3

# Scheduler / background
apscheduler==3.10.4

# Env
pydantic-settings==2.6.1
